pyproject.toml
[project]
name = "phone-agent"
version = "0.1.0"
description = "Phone controlling agent built using google adk"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=0.3.0",
    "google-genai>=1.10.0",
    "jinja2>=3.1.6",
    "numpy>=2.2.4",
    "Pillow>=10.0.0",
    "pyautogui>=0.9.54",
]

[tool.ruff]
select = ["E", "W", "F"]

[tool.isort]
profile = "black"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"
